move class endpoints to separate routes file
