+++ src/utils/formatters.js
@@ -8,3 +8,8 @@ export function localizeNumber(number, lang) {
     }
     return new Intl.NumberFormat(lang).format(number);
 }
+
+export const convertIfNumber = (str) => {
+    const trimmed = str.trim();
+    return !isNaN(trimmed) && trimmed !== '' ? Number(trimmed) : str;
+};
